Ensure clusterio processer are shutdown
Take extra steps to make sure the clusterio master and slave spawned
during testing is really terminated when the test environment exits.

+++ test/teardown.ts
@@ -2,13 +2,19 @@ import child_process from "child_process";
 import events from "events";
 
 declare namespace global {
-	var masterProcess: child_process.ChildProcess;
-	var slaveProcess: child_process.ChildProcess;
+	var masterProcess: child_process.ChildProcess | undefined;
+	var slaveProcess: child_process.ChildProcess | undefined;
 }
 
 export default async function setup() {
-	global.slaveProcess.kill("SIGINT");
-	await events.once(global.slaveProcess, "exit");
-	global.masterProcess.kill("SIGINT");
-	await events.once(global.masterProcess, "exit");
+	if (global.slaveProcess) {
+		global.slaveProcess.kill("SIGINT");
+		await events.once(global.slaveProcess, "exit");
+		delete global.slaveProcess
+	}
+	if (global.masterProcess) {
+		global.masterProcess.kill("SIGINT");
+		await events.once(global.masterProcess, "exit");
+		delete global.masterProcess;
+	}
 }
